get challenges api and persist submission data

+++ src/main/java/vn/candicode/payload/response/SubmissionSummary.java
@@ -11,8 +11,8 @@ import java.util.List;
 @Setter
 @Builder
 public class SubmissionSummary implements Serializable {
-    private Long total; // number of testcases
-    private Long passed; // number of passed testcases
+    private Integer total; // number of testcases
+    private Integer passed; // number of passed testcases
     private String compiled; // compile status, 'success' or 'failed'
     private String error; // compile error in detail
     private List<SubmissionDetails> details; // detailed results for each testcases, empty if failed to compile
